[SID:3604] fix: test_3502_insights_board.py 「경합 시 피해자」 근본원인 분리 — app.main import 1회 비용을 HTTP 라우터 6테스트로 격리
--durations=0으로 원 파일(20개 테스트) 프로파일 — test_get_insights_board_
endpoint_agent_200_with_rows 1건만 4.08s(나머지 19개는 전부 0.3~0.4s). 원인:
이 테스트가 파일 안에서 `from app.main import app`을 처음 호출하는 자리라
FastAPI 앱 전체 import(모든 라우터·서비스·모델 등록)를 1회 지불한다 —
`uv run python3 -c "import app.main"` 격리 측정으로 정확히 3.387s 확認.

CI는 destructive_schema 파일마다 완전히 새 `uv run pytest` 프로세스를 띄운다
(ci.yml "isolated fresh DB per file, this shard" 루프 — 파일 간 프로세스
공유 0). 즉 이 3.4s는 이 파일의 HTTP 엔드포인트 테스트(6개, `from app.main
import app` 필요) 쪽에서만 발생하고, 나머지 14개(list_insights_board를
서비스 함수로 직접 호출, app.main 무관)는 전혀 안 낸다 — 그런데 등재 39s는
20개를 한 파일로 뭉쳐 쟀으므로 이 CPU-bound 임포트 비용이 나머지 I/O-bound
14개 테스트의 안정성까지 같이 흔들었다(경합 시 80s로 튄 관측과 부합 — CPU
경합은 CPU-bound 작업을 I/O-bound 작업보다 훨씬 나쁘게 저하시킨다).

story #3502의 원 파일 자체가 이미 "조각①(서비스 함수)·조각②(HTTP 라우터)"로
주석 구획돼 있었다(§649) — 그 경계 그대로 물리 분할:
- test_3502_insights_board.py: 조각①만 남김(14개, list_insights_board 직접
  호출, app.main 무관). 격리 uv run pytest 2회 로컬 실측 6.65~6.68s.
- test_3502_insights_board_endpoints.py(신규): 조각②(6개, app.main 1회
  비용 포함). 격리 uv run pytest 2회 로컬 실측 4.63~4.85s.
둘 다 20s 목표 아래(경합 배율 2배 감안해도 여유) — 테스트 수(14+6=20)·
assertion 전부 원본 그대로, 세팅 헬퍼(_seed_site_post 등)만 필요한 쪽에 복제
(test_3471_org_content_rules_lint.py 공용 헬퍼는 기존과 동일하게 재사용).

infra/destructive-schema-shard-weights.json — 두 항목 각각 로컬 실측×2배
안전마진 잠정값(14s·10s)으로 등재(source에 방법론 명시, 실 CI 샤드 로그로
재확認 필요 표시). lint_destructive_schema_weights_registered.py·
lint_destructive_publish_path_dispose_fixture.py·lint_destructive_test_sql.py
로컬 재실행 확認(전부 OK).
