+++ backend/tests/test_3502_insights_board.py
@@ -645,182 +645,8 @@ async def test_invalid_window_raises():
     finally:
         await engine.dispose()
 
-
-# ─── 조각②: HTTP 라우터(GET insights-board · POST follow-ups) ─────────────────
-
-
-@pytest.mark.anyio
-async def test_get_insights_board_endpoint_agent_200_with_rows():
-    from app.main import app
-
-    engine, Session = await _session_factory()
-    try:
-        async with Session() as s:
-            org_id, project_id = await _seed_org(s)
-            agent_id = await _seed_agent(s, org_id, project_id)
-            story_id = await _seed_story(s, org_id, project_id)
-            await _seed_site_post(
-                s, org_id=org_id, work_item_id=story_id, slug="post-http", title="HTTP",
-                published_at=datetime.now(timezone.utc) - timedelta(days=1),
-            )
-
-        _setup_org_scoped_app(app, Session, org_id, user_id=agent_id, agent=True)
-        async with _client_for(app) as client:
-            r = await client.get(f"/api/v2/organizations/{org_id}/insights-board")
-        assert r.status_code == 200, r.text
-        body = r.json()
-        assert len(body["rows"]) == 1
-        assert body["rows"][0]["title"] == "HTTP"
-        assert body["has_more"] is False
-    finally:
-        app.dependency_overrides.clear()
-        await engine.dispose()
-
-
-@pytest.mark.anyio
-async def test_get_insights_board_endpoint_invalid_window_422():
-    from app.main import app
-
-    engine, Session = await _session_factory()
-    try:
-        async with Session() as s:
-            org_id, project_id = await _seed_org(s)
-            agent_id = await _seed_agent(s, org_id, project_id)
-
-        _setup_org_scoped_app(app, Session, org_id, user_id=agent_id, agent=True)
-        async with _client_for(app) as client:
-            r = await client.get(f"/api/v2/organizations/{org_id}/insights-board?window=14d")
-        assert r.status_code == 422, r.text
-        assert r.json()["error"]["code"] == "INSIGHTS_BOARD_INVALID_WINDOW"
-    finally:
-        app.dependency_overrides.clear()
-        await engine.dispose()
-
-
-@pytest.mark.anyio
-async def test_create_follow_up_creates_story_with_number_and_evidence():
-    from app.main import app
-    from app.models.evidence import Evidence
-    from app.models.pm import Story
-    from sqlalchemy import select
-
-    engine, Session = await _session_factory()
-    try:
-        async with Session() as s:
-            org_id, project_id = await _seed_org(s)
-            human_id = await _seed_human(s, org_id, role="member")
-            story_id = await _seed_story(s, org_id, project_id, title="원문")
-            sp = await _seed_site_post(
-                s, org_id=org_id, work_item_id=story_id, slug="post-fu", title="FU",
-                published_at=datetime.now(timezone.utc) - timedelta(days=2),
-            )
-
-        _setup_org_scoped_app(app, Session, org_id, user_id=human_id)
-        async with _client_for(app) as client:
-            r = await client.post(
-                f"/api/v2/organizations/{org_id}/publications/{sp.id}/follow-ups",
-                json={"kind": "republish", "note": "다시 내보내자"},
-            )
-        assert r.status_code == 201, r.text
-        new_story_id = uuid.UUID(r.json()["story_id"])
-
-        async with Session() as s:
-            new_story = (await s.execute(select(Story).where(Story.id == new_story_id))).scalar_one()
-            assert new_story.project_id == project_id
-            assert new_story.story_number is not None, "allocate_story_number()가 채번해야 한다"
-            assert "원문" in new_story.title
-
-            evidence = (await s.execute(
-                select(Evidence).where(Evidence.work_item_id == new_story_id)
-            )).scalar_one()
-            assert evidence.payload["kind"] == "follow_up_created"
-            assert evidence.payload["follow_up_kind"] == "republish"
-            assert evidence.payload["publication_id"] == str(sp.id)
-            assert evidence.payload["recorded_by"] == "platform"
-            assert evidence.created_by is None
-    finally:
-        app.dependency_overrides.clear()
-        await engine.dispose()
-
-
-@pytest.mark.anyio
-async def test_create_follow_up_agent_forbidden():
-    from app.main import app
-
-    engine, Session = await _session_factory()
-    try:
-        async with Session() as s:
-            org_id, project_id = await _seed_org(s)
-            agent_id = await _seed_agent(s, org_id, project_id)
-            story_id = await _seed_story(s, org_id, project_id)
-            sp = await _seed_site_post(
-                s, org_id=org_id, work_item_id=story_id, slug="post-fu-agent", title="FUA",
-                published_at=datetime.now(timezone.utc) - timedelta(days=2),
-            )
-
-        _setup_org_scoped_app(app, Session, org_id, user_id=agent_id, agent=True)
-        async with _client_for(app) as client:
-            r = await client.post(
-                f"/api/v2/organizations/{org_id}/publications/{sp.id}/follow-ups",
-                json={"kind": "edit"},
-            )
-        assert r.status_code == 403, r.text
-        assert r.json()["error"]["code"] == "FOLLOW_UP_CREATE_HUMAN_ONLY"
-    finally:
-        app.dependency_overrides.clear()
-        await engine.dispose()
-
-
-@pytest.mark.anyio
-async def test_create_follow_up_other_org_publication_404():
-    from app.main import app
-
-    engine, Session = await _session_factory()
-    try:
-        async with Session() as s:
-            org_a, project_a = await _seed_org(s)
-            story_a = await _seed_story(s, org_a, project_a)
-            sp_a = await _seed_site_post(
-                s, org_id=org_a, work_item_id=story_a, slug="post-org-a", title="A",
-                published_at=datetime.now(timezone.utc) - timedelta(days=1),
-            )
-            org_b, project_b = await _seed_org(s)
-            human_b = await _seed_human(s, org_b, role="member")
-
-        _setup_org_scoped_app(app, Session, org_b, user_id=human_b)
-        async with _client_for(app) as client:
-            r = await client.post(
-                f"/api/v2/organizations/{org_b}/publications/{sp_a.id}/follow-ups",
-                json={"kind": "stop"},
-            )
-        assert r.status_code == 404, r.text
-    finally:
-        app.dependency_overrides.clear()
-        await engine.dispose()
-
-
-@pytest.mark.anyio
-async def test_create_follow_up_invalid_kind_422():
-    from app.main import app
-
-    engine, Session = await _session_factory()
-    try:
-        async with Session() as s:
-            org_id, project_id = await _seed_org(s)
-            human_id = await _seed_human(s, org_id, role="member")
-            story_id = await _seed_story(s, org_id, project_id)
-            sp = await _seed_site_post(
-                s, org_id=org_id, work_item_id=story_id, slug="post-fu-badkind", title="Bad",
-                published_at=datetime.now(timezone.utc) - timedelta(days=1),
-            )
-
-        _setup_org_scoped_app(app, Session, org_id, user_id=human_id)
-        async with _client_for(app) as client:
-            r = await client.post(
-                f"/api/v2/organizations/{org_id}/publications/{sp.id}/follow-ups",
-                json={"kind": "delete_everything"},
-            )
-        assert r.status_code == 422, r.text
-    finally:
-        app.dependency_overrides.clear()
-        await engine.dispose()
+# ─── 조각②(HTTP 라우터: GET insights-board · POST follow-ups)는 story #3604로
+# test_3502_insights_board_endpoints.py에 분리됐다 — `from app.main import app`
+# 1회 비용(~3.4s, 로컬 무경합)이 이 파일에 섞여 있을 때 60초 러너 가드의 등재/경합
+# 배율을 왜곡했다(그 파일 머리 주석에 실측 상세). 이 파일에 남은 14개는 전부
+# list_insights_board를 서비스 함수로 직접 부르며 app.main을 안 건드린다.
